Pufferfish- final texture!
All textures currently released in the game, including snapshots, are textured! Now that everything is textured, I can focus more on improving the quality of some of the older textures in the pack. This marks release 1.0. I will probably take a break on texturing for a bit, and focus more on developing the shader.