Shaders- stable, but rainbows
The old shaders do cause a computer crash, but the blurred screen was just from a borked nvidia config file. These new shaders calculate an average using an lod calculation. Unfortunately, it looks like the lod only samples from a very small space, so it flickers and flashes. I also need to desaturate the average, which would fix the rainbowing, but this is interesting so I thought I'd commit and share.