Semi-functional shaders!
Ok, they're ridiculously buggy, but the concept works. With a greater number of samples, the screen should not flicker. With a polynomial weighting curve, the desaturation will not affect daylight. Maybe another pass would be good to curve the sat levels for saturation weighting per pixel. But the screen *does* desaturate based on the lightness of the samples.