Shaders: tuned!
Ok, I'm still going to work on these more, but I'm going to take a break for a few days. I'd like to build out more passes for the day, in particular, and tweak some values. Bloom could be a bit stronger in the day. The bloom, screen averager and desaturation is working at this time. I will be working on the blur to optimize it, but I'm pretty happy with my little regular distribution function and don't want to mess it up. :P