Shaders: partway there...
Quite a few changes here... overall things don't work as well as they used to, but it can be fixed. It runs faster though. And all the code is prettier. Blur now takes 18 samples instead of 64 samples. Bloom gets smoothed. Top json is restructured, and the whole pipeline is arranged differently now. I'm need to re-implement the split night/day system, but do it in separate passes this time. Then use it to add a curves system and re-add night desat.